// src/shared/components/modal/modal.types.ts
export type ModalSize = 'sm' | 'md' | 'lg' | 'xl' | 'full'

export type HeaderVariant =
  | 'close-only'
  | 'back-close'
  | 'back-only'
  | 'actions'
  | 'none'

export type FooterVariant =
  | 'confirm-cancel'
  | 'ok-only'
  | 'close-only'
  | 'delete-cancel'
  | 'custom'

export type Intent = 'default' | 'danger' | 'warning' | 'info'

export type FooterIntent = 'primary' | 'danger' | 'warning'

export type Align = 'right' | 'left' | 'center' | 'between'

export type Padding = 'none' | 'sm' | 'md' | 'lg'